[Refactor configuration] Refactor configuration files of environment variables for smoother and cleaner app
