+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UserModule } from './user/user.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { typeOrmAsyncConfig } from 'database/data-source';
+import { typeOrmAsyncConfig } from 'database/type.orm.async.config';
 import { ConfigModule } from '@nestjs/config';
 import configuration from './config/configuration';
 require('dotenv').config()
@@ -11,7 +11,7 @@ require('dotenv').config()
 @Module({
   imports: [
     ConfigModule.forRoot({
-      envFilePath: [`.env.${process.env.NODE_ENV}`],
+      envFilePath: [`${process.cwd()}/.env.${process.env.NODE_ENV}`],
       isGlobal: true,
       load: [configuration],
     }),
