+++ src/main.ts
@@ -4,9 +4,11 @@ import { ConfigService } from '@nestjs/config';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+
   const configService = app.get(ConfigService);
+  
   const port = configService.get<number>('port');
   await app.listen(port);
-  console.log(port)
+  console.log(`This app is running on port ${port}`)
 }
 bootstrap();
